% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.slasso_cv}
\alias{plot.slasso_cv}
\alias{plot.slasso}
\title{Plot the results of  the S-LASSO method}
\usage{
\method{plot}{slasso_cv}(mod, type = "", ss_rule_par = 1, fun_c = "mean", ylim = NA, ...)

\method{plot}{slasso}(mod, ...)
}
\arguments{
\item{...}{No additional parameters, called for side effects.}

\item{x}{The output of  either `slasso.fr_cv` or `slasso.fr`.}
}
\value{
No return value, called for side effects.
}
\description{
This function provides plots of the estimated cluster mean functions and of the classified curves when applied to the output of `sasfclust`, whereas
 provides the cross-validation plots when applied to the output of `sasfclust_cv`. In the latter case the first plot displays the CV values as a function of  \code{G}, \code{lambda_s} and \code{lambda_l};
 the second plot displays the CV values as a function of \code{lambda_s} and \code{lambda_l} for \code{G} fixed at its optimal value;
  the third plot displays the CV values as a function of \code{lambda_l} for \code{G} and \code{lambda_s}   fixed at their optimal value.
}
\examples{
\donttest{
library(slasso)

}
}
