\name{slasso-package}
\alias{slasso-package}
\alias{slasso}
\docType{package}
\title{
\packageTitle{slasso}
}
\description{
\packageDescription{slasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{slasso}
\packageIndices{slasso}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{slasso}

Maintainer: \packageMaintainer{slasso}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
