% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.slasso}
\alias{plot.slasso}
\alias{plot.slasso_cv}
\title{Plot the results of  the S-LASSO method}
\usage{
\method{plot}{slasso_cv}(x, ...)

\method{plot}{slasso}(x, ...)
}
\arguments{
\item{x}{The output of  either \code{slasso.fr_cv} or \code{slasso.fr}.}

\item{...}{No additional parameters, called for side effects.}
}
\value{
No return value, called for side effects.
}
\description{
This function provides plots of the S-LASSO coefficient function estimate when applied to the output of \code{slasso.fr}, whereas
provides the cross-validation plots when applied to the output of \code{slasso.fr_cv}. In the latter case the first plot displays the CV values as a function of  \code{lambda_L}, \code{lambda_s} and \code{lambda_t}, and
the second plot displays the CV values as a function of \code{lambda_L} with \code{lambda_s} and \code{lambda_t} fixed at their optimal values.
}
\examples{
\donttest{
library(slasso)

}
}
