% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slasso-package.R
\docType{package}
\name{slasso-package}
\alias{slasso-package}
\alias{_PACKAGE}
\alias{{rofanova}-package}
\title{slasso}
\description{
Implements the S-LASSO (RoFANOVA), described in Centofanti et al. (2021) <arXiv:>. It allows testing mean differences among groups of functional data by being robust aginst the presence of outliers.
}
\details{
\tabular{ll}{
Package: \tab rofanova\cr
Type: \tab Package\cr
Version: \tab `r packageVersion("rofanova")` \cr
Date: \tab  `r Sys.Date()` \cr
License: \tab `r packageDescription("rofanova", fields="License")`\cr
}
}
\examples{
library(rofanova)
library(rofanova)
data_out<-simulate_data(scenario="one-way")
label_1=data_out$label_1
X_fdata<-data_out$X_fdata
B=10
cores=1
per_list_median<-rofanova(X_fdata,label_1,B = B,family="median",cores=cores)
pvalue_median_vec<-per_list_median$pval_vec
per_list_huber<-rofanova(X_fdata,label_1,B = B,family="huber",cores=cores)
pvalue_huber_vec<-per_list_huber$pval_vec
per_list_bisquare<-rofanova(X_fdata,label_1,B = B,family="bisquare",cores=cores)
pvalue_bisquare_vec<-per_list_bisquare$pval_vec
per_list_hampel<-rofanova(X_fdata,label_1,B = B,family="hampel",cores=cores)
pvalue_hampel_vec<-per_list_hampel$pval_vec
per_list_optimal<-rofanova(X_fdata,label_1,B = B,family="optimal",cores=cores)
pvalue_optimal<-per_list_optimal$pval
}
\references{
Centofanti, F., Lepore, A., & Palumbo, B. (2021).
}
\seealso{
\code{\link{fusem}} \code{\link{funmad}}
}
\author{
Fabio Centofanti, Antonio Lepore, Biagio Palumbo
}
